refactor(firestore-translate-text): improve readability and testability
