+++ firestore-translate-text/functions/jest.config.js
@@ -11,7 +11,7 @@ module.exports = {
   },
   preset: "ts-jest",
   setupFiles: ["<rootDir>/__tests__/jest.setup.ts"],
-  testMatch: ["**/__tests__/*.test.ts"],
+  testMatch: ["**/__tests__/unit/translateMultipleBackfill.test.ts"],
   moduleNameMapper: {
     "firebase-admin/eventarc":
       "<rootDir>/node_modules/firebase-admin/lib/eventarc",
