+++ firestore-translate-text/functions/package.json
@@ -12,20 +12,20 @@
     "generate-readme": "firebase ext:info .. --markdown > ../README.md"
   },
   "dependencies": {
-    "@types/node": "14.18.34",
     "@google-cloud/translate": "^4.0.1",
+    "@types/express-serve-static-core": "4.17.30",
+    "@types/node": "14.18.34",
     "firebase-admin": "^11.4.1",
     "firebase-functions": "^3.24.0",
-    "typescript": "^4.4.4",
     "rimraf": "^2.6.3",
-    "@types/express-serve-static-core": "4.17.30"
+    "typescript": "^4.4.4"
   },
   "devDependencies": {
+    "@types/jest": "^24.0.18",
     "firebase-functions-test": "^0.1.7",
+    "jest": "^24.9.0",
     "js-yaml": "^3.13.1",
     "mocked-env": "^1.3.1",
-    "@types/jest": "^24.0.18",
-    "jest": "^24.9.0",
     "ts-jest": "^24.1.0"
   },
   "private": true
