+++ firestore-translate-text/functions/src/translate/index.ts
@@ -1,0 +1,6 @@
+export {
+  translateDocument,
+  translateDocumentBackfill,
+} from "./translateDocument";
+
+export { updateTranslations, extractLanguages, extractInput } from "./common";
